fix(tests): update player ID type in disambiguation display tests
- Changed player ID type from `usize` to `i64` in the disambiguation display tests to ensure consistency and prevent potential overflow issues with larger datasets.
- This modification enhances the robustness of the tests by aligning the player ID type with expected data structures.

These changes aim to improve the accuracy and reliability of the disambiguation display tests.

+++ tests/disambiguation_display_tests.rs
@@ -785,12 +785,12 @@ fn test_disambiguation_performance_with_many_players() {
     // Create a large set of players that need disambiguation
     let mut home_players = Vec::new();
     for i in 0..50 {
-        home_players.push((i, format!("Player{i}"), "Koivu".to_string()));
+        home_players.push((i as i64, format!("Player{i}"), "Koivu".to_string()));
     }
 
     let mut away_players = Vec::new();
     for i in 50..100 {
-        away_players.push((i, format!("Player{i}"), "Selänne".to_string()));
+        away_players.push((i as i64, format!("Player{i}"), "Selänne".to_string()));
     }
 
     // Create disambiguation contexts for testing
